Renamed file newexercisedemandhandler. Reworked tests relating to refactoring.
