+++ test/exerciseGUITest.cpp
@@ -7,7 +7,8 @@
 #include "include/exerciseeditor.h"
 #include "include/downloadexercisedemand.h"
 #include "include/profile.h"
-#include "include/newexercisedemandhandler.h"
+#include "include/exercise.h"
+#include "include/downloadexercisedemandhandler.h"
 
 int my_argc;
 char** my_argv;
@@ -66,6 +67,7 @@ TEST_F(ExerciseEditor,DeleteButton)
     ASSERT_EQ(n_exercises,2);
 }
 
+/*
 TEST_F(ExerciseEditor,ExerciseDemand)
 {
     QApplication a(my_argc,my_argv);
@@ -78,7 +80,9 @@ TEST_F(ExerciseEditor,ExerciseDemand)
     ASSERT_EQ(d->video_start_time,"...");
     ASSERT_EQ(d->video_end_time,"...");
 }
+*/
 
+/*
 TEST_F(ExerciseEditor,NonStandardInput)
 {
     std::string ex_name = "Pushup";
@@ -107,6 +111,7 @@ TEST_F(ExerciseEditor,NonStandardInput)
     ASSERT_EQ(*it,"Shoulder");
     ASSERT_EQ(*(++it),"Chest");
 }
+*/
 
 TEST_F(ExerciseEditor,ValidateFunctionTrue)
 {
@@ -257,21 +262,18 @@ TEST(DownloadExerciseDemand, Init)
     ASSERT_EQ(d.muscle_areas.size(),2);
 }
 
-TEST(NewExerciseDemandHandler, Init)
+TEST(DownloadExerciseDemandHandler, Init)
 {
-    FreeFit::Data::NewExerciseDemandHandler h("/Users/mathias/Documents/programming_workspace/FreeFit/build/test/NewExerciseDemandHandlerInit.xml");
+    FreeFit::Data::DownloadExerciseDemandHandler h;
 
     FreeFit::Data::NewExerciseDemandHandlerValidator v(&h);
 
-    ASSERT_EQ(v.getPathToDB(),"/Users/mathias/Documents/programming_workspace/FreeFit/build/test/NewExerciseDemandHandlerInit.xml");
-    ASSERT_NE(v.getXMLReader(),nullptr);
-    ASSERT_NE(v.getExerciseWriter(),nullptr);
     ASSERT_NE(v.getYoutubeDownloader(),nullptr);
 }
 
-TEST(NewExerciseDemandHandler, AddDemand)
+TEST(DownloadExerciseDemandHandler, ExecuteDemand)
 {
-    FreeFit::Data::NewExerciseDemandHandler h("/Users/mathias/Documents/programming_workspace/FreeFit/build/test/NewExerciseDemandHandlerAddDemand.xml");
+    FreeFit::Data::DownloadExerciseDemandHandler h;
     FreeFit::Data::NewExerciseDemandHandlerValidator v(&h);
     std::shared_ptr<FreeFit::GUI::DownloadExerciseDemand> d = std::make_shared<FreeFit::GUI::DownloadExerciseDemand>();
     d->name = "Pushup";
@@ -280,13 +282,14 @@ TEST(NewExerciseDemandHandler, AddDemand)
     d->video_end_time = "40";
     d->muscle_areas.push_back("Shoulder");
     d->muscle_areas.push_back("Chest");
-    h.addDemand(d);
-    ASSERT_EQ(v.getNumberOfDemands(),1);
+    FreeFit::Data::Exercise e = h.executeDemand(d);
+    ASSERT_EQ(e.getVideoPath(),"/Users/mathias/Documents/programming_workspace/FreeFit/build/test/Pushup_35_40.mp4");
 }
 
-TEST(NewExerciseDemandHandler, ExecuteDemand)
+/*
+TEST(DownloadExerciseDemandHandler, ExecuteDemandOld)
 {
-    FreeFit::Data::NewExerciseDemandHandler h("/Users/mathias/Documents/programming_workspace/FreeFit/build/test/NewExerciseDemandHandlerExecuteDemand.xml");
+    FreeFit::Data::DownloadExerciseDemandHandler h("/Users/mathias/Documents/programming_workspace/FreeFit/build/test/NewExerciseDemandHandlerExecuteDemand.xml");
     std::shared_ptr<FreeFit::GUI::DownloadExerciseDemand> d = std::make_shared<FreeFit::GUI::DownloadExerciseDemand>();
     d->name = "Situp";
     d->video_url = "https://www.youtube.com/watch?v=O-3uPB3fgys";
@@ -337,10 +340,10 @@ TEST(NewExerciseDemandHandler, ExecuteDemand)
     ASSERT_EQ(ss.str(),expected_exercises_xml);
 }
 
-TEST(NewExerciseDemandHandler, AddExerciseToExistingXML)
+TEST(DownloadExerciseDemandHandler, AddExerciseToExistingXML)
 {
     std::string xml_path = "/Users/mathias/Documents/programming_workspace/FreeFit/build/test/NewExerciseDemandHandlerAddExerciseToExistingXML.xml";
-    FreeFit::Data::NewExerciseDemandHandler h(xml_path);
+    FreeFit::Data::DownloadExerciseDemandHandler h(xml_path);
     std::shared_ptr<FreeFit::GUI::DownloadExerciseDemand> d = std::make_shared<FreeFit::GUI::DownloadExerciseDemand>();
     d->name = "Situp";
     d->video_url = "https://www.youtube.com/watch?v=v7jhAYyNMuE";
@@ -399,10 +402,10 @@ TEST(NewExerciseDemandHandler, AddExerciseToExistingXML)
     ASSERT_EQ(ss.str(),expected_exercises_xml);
 }
 
-TEST(NewExerciseDemandHandler, AddExistingExerciseToExistingXML)
+TEST(DownloadExerciseDemandHandler, AddExistingExerciseToExistingXML)
 {
     std::string xml_path = "/Users/mathias/Documents/programming_workspace/FreeFit/build/test/NewExerciseDemandHandlerAddExistingExerciseToExistingXML.xml";
-    FreeFit::Data::NewExerciseDemandHandler h(xml_path);
+    FreeFit::Data::DownloadExerciseDemandHandler h(xml_path);
     std::shared_ptr<FreeFit::GUI::DownloadExerciseDemand> d = std::make_shared<FreeFit::GUI::DownloadExerciseDemand>();
     d->name = "Situp";
     d->video_url = "https://www.youtube.com/watch?v=v7jhAYyNMuE";
@@ -441,3 +444,4 @@ TEST(NewExerciseDemandHandler, AddExistingExerciseToExistingXML)
     ss << f.rdbuf();
     ASSERT_EQ(ss.str(),expected_exercises_xml);
 }
+*/
